+++ CHANGELOG.md
@@ -5,6 +5,9 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Removed
+- file-storage: remove `Shares` from `ModifyFileStorageRequest` and `ReplaceFileStorageRequest`
+
 ## [8.31.1]
 
 ### Changed